Working towards GPU-ness
Working on folding the mass_attenuation Python code into the projectKernel function.  What I have NOT yet finished for this "folding" is implemented the get_absorbtion_coefs(energy, mat) call in CUDA.

Added mat_coefs_for_kernel.cu to hold constant material_coefficient values, but won't need that file later.

The insight is that get_absorbtion_coefs(energy, mat) is NOT a function of the pixel--accordingly, we can pre-compute the values and store them in a [n_bins x NUM_MATERIALS] table, where n_bins is the number of different energy values ("spectral bins").  This prec-computed table will be a field of the projector.py:Projector class, and can be passed in as an argument to project_kernel.cu:project_kernel(...) via the projector.py:Projector._project(self, ...) method.