very interested pdf normalization behavior
Whether or not I normalize noncont_pdf (by dividing it by its sum), the output image is the same.  Accordingly, the diff image is the same whether or not I normalize the pdf.

That the code behaves the same if pdf is changed by a factor f_1 versus f_2 is very interesting.  Could that be the source of my bug?