Refactored RITA and figured out how to use it
Turned the RITA parameter arrays into a class, rita.py:RITA.  This allows for a dictionary, Dict[str, RITA] that maps the material (e.g. "bone", "soft tissue") to the RITA sampler.

Based on my reading of the PENELOPE-2006 specifications, the distribution we are sampling, \pi(x^2), is distributed like [F(x,Z)]^2.  However, for a given photon with dimensionless energy \kappa, we only want to sample values on the interval (0, (20.6074 * 2\kappa)^2).  So, it appears that a RITA object is a function of both the material AND the energy of the incoming photon, since a RITA object is specified by the PDF to approximate as well as the interval to approximate over.  This dependency on material AND energy is bad, since we don't want to recalculate a RITA object for every single photon energy level within a material.

The way around this need for recalculation is to rely on resampling.  We initialize a sinlge RITA object for each material, based on the maximum possible energy of a photon from the spectrum that we are using.  Then, when sampling for photons with lower energy levels, we reject any RITA samples that fall outside the desired interval _for_that_photon_.