Fixing __init__.py syntax and initial data dump
In __init__.py, __all__ is supposed to be a list of strings (https://docs.python.org/3/tutorial/modules.html).

Initial (and partial) data dump of Mean Free Path data for scatter simulation.  Mainly a proof of concept.  Will be updating shortly to have all of the data (which is A LOT).

Issues with the MFP data: I do not currently have data corresponding to the following materials (see material_coefficients.py):
- soft tissue
- iron
- teflon
- anything called "external"