Finally got a clean solution for the detector plane issue
Figured out a clean geometric solution to representing the 'plane vector' of the detector plane, complete with thorough comments.  Is very much better than the "analytically find the global minimum of a function of two variables" method that I was trying to use for the "distance from the origin" part of the detector plane vector.

Still some TODOs before implementing variance reduction techniques, namely:
1. Math for determining which pixel got hit, in function track_single_photon_no_vr(...)
2. Converting the photon energies into image intensities in function simulate_scatter_no_vr(...)