Toward finishing initialization tracking
Also fixed some stuff with the interaction-type sampling in Python.

Currently having trouble combining the mass-attenuated primary image with the scatter image; the background color of the final image doesn't seem the shade that it should be.

I think the reason that combining the image and noise is faulty is because the mass-attenuation code and the scatter code are returning different quantities.  In project_kernel(...), deposited_energy[...] has units of eV, BUT SCALED by 1) the attenuation and 2) the pdf for each energy's spectral bin.  To achieve the ACTUAL value of the deposited energy, I think the currently-returned value needs to be multiplied by the photon count used in the projection.

Once I have taken that step, I think the primary and scatter should mesh if the photon count used for both is equal.  That will be something to investigate.