Almost wrapped the CUDA scatter
A bit more work to be done to have PyCuda properly call the scatter kernel.

Additionally, I still need to verify the math for combining the scatter signal with the primary signal.

I think I should return the 'intensity' output to being [deposited energy per photon].  That output would not be distorted by the differences in solid angle between the pixels, and thus is useful for the intensity_upper_bound clipping.