Default behavior returns to 'intensity'
I implemented the collected-energy conversion logic.

Something that would be quite nice would be to intelligibly return the scatter signal as some sort of "signal per photon" value.  Would it be "signal per simulated photon"? Would it be "signal per detected scattered photon" (I think so)?  Thinking about this is a TODO