More bugs
When splicing volumes of the same material, the boundary between them it pretty obvious, even when it should be invisible.  For example, consider the following two-volume scenario:
  ____________
/___/___/___/|
|___|___ |___| |
|___|___ |___| |
|___|___ |___|/

and we define the volumes:
- volume0: all boxes, all soft tissue
- volume1: the central box, all soft tissue

The boundary between the volumes appears darker than the surrounding pixels, even though the materials are the same.  This happens regardless of whether the input is:
- [(volume0, 0), (volume1, 0)], or
- [(volume0, 1), (volume1, 0)]