Outsourcing lps_from_ijk
Using nibabel.nicom.dicomwrappers.MultiframeWrapper to
construct the affine transform from the available dicom
tags.

Evidently this is complex enough to _not_ maintain it
in this project. See this post to get a feeling for the
complexity: https://nipy.org/nibabel/dicom/dicom_orientation.html#defining-the-dicom-orientation.
The used class warns, that the exposed functionality
is not well tested. I will verify that it works with
our datasets.